small updates
(tao)
%
Subtlety

The ancients did not seek to rule people with knowledge,
But to help them become natural.
It is difficult for knowledgeable people to become natural.
To use law to control a nation weakens the nation.
But to use nature to control a nation strengthens the nation.
Understanding these two paths is understanding subtlety;
Subtlety runs deep, ranges wide,
Resolves confusion and preserves peace.
		-- Lao Tse, "Tao Te Ching"